QnAGenie/requirements.txt
fastapi
uvicorn
python-multipart
openai
pandas
python-dotenv
pdfplumber
pytesseract
Pillow
python-jose[cryptography]